+++ 1.21.6/Common/src/main/java/fuzs/puzzleslib/impl/core/proxy/EntityProxy.java
@@ -10,7 +10,7 @@ public interface EntityProxy {
 
     boolean canEquip(ItemStack itemStack, EquipmentSlot equipmentSlot, LivingEntity livingEntity);
 
-    @Nullable EntitySpawnReason getMobSpawnType(Mob mob);
+    @Nullable EntitySpawnReason getMobSpawnReason(Mob mob);
 
     boolean isMobGriefingAllowed(ServerLevel serverLevel, @Nullable Entity entity);